Basic functionality is now working. You can zoom, load POIs and click them to see more.
Currently it fetches only data from OpenStreetMap but since there are now some shelters and campsites being added to Wikidata the plan is to support that too in the next version.

It is probably feasible to evolve this application into using some kind of database and load from there dynamically but that is more complex as it requires updating the database regularly.

The current map is Thunderforest Outdoors that shows hiking routes but unfortunally it does not seem to show mixed routes (hiking;running) or other routes that hiking, e.g. canoo routes or riding routes.